ACA-68 - move ChaneOrganisationRequest class to correct `domain` package
