+++ src/main/java/uk/gov/hmcts/reform/managecase/domain/ChangeOrganisationRequest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.managecase.client.datastore.model;
+package uk.gov.hmcts.reform.managecase.domain;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
